+++ CHANGELOG.md
@@ -7,10 +7,12 @@
 - Return the effective `setup_duration` in the solution as `setup_time` [#428](https://github.com/Mapotempo/optimizer-api/pull/428) [#445](https://github.com/Mapotempo/optimizer-api/pull/445)
 - Exclusive position for activities that needs to be both first and last position [#438](https://github.com/Mapotempo/optimizer-api/pull/438)
 - The intermediate solutions are now cleaned similarly to the final solution [#435](https://github.com/Mapotempo/optimizer-api/pull/435)
+- Now VROOM can handle setup_duration [#436](https://github.com/Mapotempo/optimizer-api/pull/436/)
 
 ### Changed
 
 - Base `setup_duration` within optimizer-ortools on `point_id` instead of `matrix_index` [#444] (https://github.com/Mapotempo/optimizer-api/pull/444) [#446](https://github.com/Mapotempo/optimizer-api/pull/446]
+- Bump VROOM to v1.12.0 [#436](https://github.com/Mapotempo/optimizer-api/pull/436/)
 
 ### Removed